Move SlotEntities from models.py to websocket.py
Same audit principle as SlotState/SlotMetadata: SlotEntities is
websocket-only — used by _get_slot_entity_ids and _get_slot_entity_data
inside websocket.py and nowhere else. Lives next to its consumers now.

models.py is now down to genuinely cross-cutting types only:
SlotCode (used by sync, providers, coordinator, data, websocket, tests)
and LockCodeManagerConfigEntryData / LockCodeManagerConfigEntry
(runtime data shape consumed by every entry-aware module via
runtime_data).

The "Canonical home for cross-cutting types" claim in models.py's
module docstring is now truthful — every type left in models.py is
actually used cross-module.

602 tests pass.

Entire-Checkpoint: f3756c1b2200
