+++ tests/test_websocket.py
@@ -64,9 +64,10 @@ from custom_components.lock_code_manager.const import (
     DOMAIN,
 )
 from custom_components.lock_code_manager.exceptions import DuplicateCodeError
-from custom_components.lock_code_manager.models import SlotCode, SlotEntities
+from custom_components.lock_code_manager.models import SlotCode
 from custom_components.lock_code_manager.providers import BaseLock
 from custom_components.lock_code_manager.websocket import (
+    SlotEntities,
     _find_config_entry_by_title,
     _get_bool_state,
     _get_condition_entity_data,
